+++ common/engine/khg_phy/include/khg_phy/constraints/spline_constraint.h
@@ -15,7 +15,7 @@ typedef struct {
   phy_vector2 anchor;
   float max_force;
   phy_vector2 controls[PHY_SPLINE_CONSTRAINT_MAX_CONTROL_POINTS];
-  size_t num_controls;
+  unsigned int num_controls;
 } phy_spline_constraint;
 
 typedef struct {
@@ -36,9 +36,9 @@ phy_vector2 phy_spline_constraint_get_anchor(const phy_constraint *cons);
 void phy_spline_constraint_set_max_force(phy_constraint *cons, float max_force);
 float phy_spline_constraint_get_max_force(const phy_constraint *cons);
 
-int phy_spline_constraint_set_control_points(phy_constraint *cons, phy_vector2 *points, size_t num_points);
+int phy_spline_constraint_set_control_points(phy_constraint *cons, phy_vector2 *points, unsigned int num_points);
 phy_vector2 *phy_spline_constraint_get_control_points(const phy_constraint *cons);
-size_t phy_spline_constraint_get_number_of_control_points(const phy_constraint *cons);
+unsigned int phy_spline_constraint_get_number_of_control_points(const phy_constraint *cons);
 
 void phy_spline_constraint_presolve(struct phy_space *space, phy_constraint *cons, float dt, float inv_dt);
 void phy_spline_constraint_warmstart(struct phy_space *space, phy_constraint *cons);
