+++ common/engine/khg_phy/src/contact.c
@@ -5,7 +5,7 @@
 
 bool phy_persistent_contact_pair_penetrating(phy_persistent_contact_pair *pcp) {
   bool penetrating = false;
-  for (size_t c = 0; c < pcp->contact_count; c++) {
+  for (unsigned int c = 0; c < pcp->contact_count; c++) {
     phy_contact contact = pcp->contacts[c];
     if (contact.separation < 0.0) {
       penetrating = true;
@@ -15,13 +15,13 @@ bool phy_persistent_contact_pair_penetrating(phy_persistent_contact_pair *pcp) {
   return penetrating;
 }
 
-uint64_t phy_persistent_contact_pair_hash(void *item) {
+unsigned long long phy_persistent_contact_pair_hash(void *item) {
   phy_persistent_contact_pair *pcp = (phy_persistent_contact_pair *)item;
   return phy_persistent_contact_pair_key(pcp->shape_a, pcp->shape_b);
 }
 
 void phy_persistent_contact_pair_remove(phy_space *space, phy_persistent_contact_pair *pcp) {
-  for (size_t c = 0; c < pcp->contact_count; c++) {
+  for (unsigned int c = 0; c < pcp->contact_count; c++) {
     phy_contact *contact = &pcp->contacts[c];
     phy_contact_event event = { .body_a = pcp->body_a, .body_b = pcp->body_b, .shape_a = pcp->shape_a, .shape_b = pcp->shape_b, .normal = pcp->normal, .penetration = contact->separation, .position = phy_vector2_add(pcp->body_a->position, contact->anchor_a), .normal_impulse = {contact->solver_info.normal_impulse}, .friction_impulse = {contact->solver_info.tangent_impulse}, .id = contact->id };
     if (space->listener && !contact->remove_invoked) {
